feat: implement environment import modal with conflict resolution and selection logic
