+++ tests/utils/page/environments/index.ts
@@ -117,12 +117,12 @@ export const buildEnvironmentLocators = (page: Page) => ({
     page.getByTestId(scope === 'global' ? 'import-global-environment-modal-submit-btn' : 'import-environment-modal-submit-btn'),
   importTotalCount: () => page.getByTestId('env-import-total-count'),
   importDuplicatesWarning: () => page.getByTestId('import-duplicates-warning'),
-  importDuplicatesGroup: () => page.locator('.group-container').filter({ hasText: 'Duplicates' }),
+  importDuplicatesGroup: () => page.getByTestId('env-import-duplicates-group'),
   importDuplicatesCount: () => page.getByTestId('env-import-duplicates-count'),
-  importNewGroup: () => page.locator('.group-container').filter({ hasText: 'New' }),
+  importNewGroup: () => page.getByTestId('env-import-new-group'),
   importNewCount: () => page.getByTestId('env-import-new-count'),
   importSelectAllCheckbox: () => page.getByTestId('env-import-select-all'),
-  importSelectedCount: () => page.locator('.footer-left-content'),
+  importSelectedCount: () => page.getByTestId('env-import-selected-count'),
   importReviewItem: (name: string) => page.getByTestId('env-import-item').filter({ has: page.getByText(name, { exact: true }) }),
   importItemCheckbox: (name: string) => buildEnvironmentLocators(page).importReviewItem(name).getByTestId('env-import-item-checkbox'),
   importCopyButton: (name: string) => buildEnvironmentLocators(page).importReviewItem(name).getByTestId('env-import-copy-btn'),
